Split OnDiskIndexBuilder from index_handlers
