+++ scripts/preprocessing/make_on_disk_index.py
@@ -23,8 +23,8 @@ arg_parser.add_option('-n', '--n_subindexes', type='int', default=-1)
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
 
-from dt_sim_api.indexer.IVF_disk_index_handler import OnDiskIndexBuilder
 from dt_sim_api.vectorizer import SentenceVectorizer
+from dt_sim_api.indexer.on_disk_index_builder import OnDiskIndexBuilder
 from dt_sim_api.processor.document_processor import DocumentProcessor
 # </editor-fold>
