+++ scripts/preprocessing/streaming_preprocessor.py
@@ -42,7 +42,7 @@ sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
 from dt_sim_api.data_reader.io_funcs \
     import check_all_docs, aggregate_all_docs, check_unique, clear
 from dt_sim_api.vectorizer.sentence_vectorizer import SentenceVectorizer
-from dt_sim_api.indexer.IVF_disk_index_handler import OnDiskIndexBuilder
+from dt_sim_api.indexer.on_disk_index_builder import OnDiskIndexBuilder
 from dt_sim_api.processor.document_processor import DocumentProcessor
 # </editor-fold>
